+++ package.json
@@ -26,13 +26,14 @@
   },
   "homepage": "https://github.com/audiojs/out#readme",
   "dependencies": {
+    "audio-buffer": "^2.4.5",
+    "audio-sink": "^1.1.6",
     "bindings": "^1.2.1",
     "debug": "^2.2.0",
     "is-audio-buffer": "^1.0.1",
     "nan": "2.2.0",
     "object-assign": "^4.1.0",
-    "pcm-util": "^2.0.3",
-    "audio-buffer": "^2.4.5"
+    "pcm-util": "^2.0.3"
   },
   "devDependencies": {
     "audio-lena": "^1.1.0"
