+++ tests/one.js
@@ -7,9 +7,11 @@ console.log('Starting test one.')
 var Speaker = AudioSpeaker({ channels: 1, float: false, bitDepth: 16, signed: true })
 var buf = new AudioBuffer(1, LenaBuffer)
 
-Speaker(buf, null, function (err, written) {
-    console.log('Test wrote ' + buf.length +' bytes of audio-lena.')
-    Speaker.end(true, (success) => {
-      console.log('Finished test one.')
-    })
+Speaker(buf, null, (err, chunk) => {
+  if(err || chunk === true) {
+    console.log('Test ended with errors.')
+  } else {
+    Speaker.end(true)
+    console.log('Test ended with no errors.')
+  }
 })
